@@ -1,12 +1,28 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.Razor">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<Nullable>disable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+		<Title>BlazorGE.Game</Title>
+		<Description>BlazorGE (Game Engine) is a lightweight Blazor game engine/framework created by Aventius and other contributors (see Git repo)</Description>
+		<PackageProjectUrl>https://github.com/aventius-software/BlazorGE</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/aventius-software/BlazorGE</RepositoryUrl>
+		<PackageTags>Blazor,aspnetcore,game-development,game-engine,game-framework,game,development,engine,framework,canvas</PackageTags>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<Version>1.0.1</Version>
+		<Company>Aventius,others</Company>
+		<Authors>Aventius,others</Authors>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\BlazorGE.Graphics2D\BlazorGE.Graphics2D.csproj" />
-    <ProjectReference Include="..\BlazorGE.Graphics\BlazorGE.Graphics.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<SupportedPlatform Include="browser" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\BlazorGE.Graphics2D\BlazorGE.Graphics2D.csproj" />
+		<ProjectReference Include="..\BlazorGE.Graphics\BlazorGE.Graphics.csproj" />
+	</ItemGroup>
 
 </Project>
