@@ -1,7 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>disable</Nullable>
+		<Title>BlazorGE.Input</Title>
+		<Description>BlazorGE (Game Engine) is a lightweight Blazor game engine/framework created by Aventius and other contributors (see Git repo)</Description>
+		<PackageProjectUrl>https://github.com/aventius-software/BlazorGE</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/aventius-software/BlazorGE</RepositoryUrl>
+		<PackageTags>Blazor,aspnetcore,game-development,game-engine,game-framework,game,development,engine,framework,canvas</PackageTags>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<Version>1.0.1</Version>
+		<Company>Aventius,others</Company>
+		<Authors>Aventius,others</Authors>
+	</PropertyGroup>
 
 </Project>
