@@ -1,5 +1,5 @@
 ﻿@inherits Canvas2DBase
 
 <div id="blazorge-canvas-container" style="opacity: 1; background-color: black; width: 100%; height: 100%" tabindex="-1">
-    <canvas id="blazorge-canvas" @ref="@CanvasReference" width="0" height="0" />    
+    <canvas id="blazorge-canvas" @ref="CanvasReference" width="1" height="1" />    
 </div>
