@@ -37,7 +37,7 @@ else
 }
 
 @code {
-    private WeatherForecast[]? forecasts;
+    private WeatherForecast[] forecasts;
 
     protected override async Task OnInitializedAsync()
     {
@@ -50,7 +50,7 @@ else
 
         public int TemperatureC { get; set; }
 
-        public string? Summary { get; set; }
+        public string Summary { get; set; }
 
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
     }
