@@ -12,5 +12,5 @@
 @code {
     // Demonstrates how a parent component can supply parameters
     [Parameter]
-    public string? Title { get; set; }
+    public string Title { get; set; }
 }
